Day 1 part 2
Instead of just adding another nested loop for part 2, I decided to try
to figure out an algorithm that would work for any group size. I'm
guessing there are more efficient ways, but this is kind of the
translation in code of the process I went through on paper to list the
possible combinations.

I decided to try to make an iterator that would yield each combination
to test one by one, which made me learn about structs, struct impls, and
trait impls for structs (if I'm even using those words correctly). I
also wanted to figure out how to organize the code so it wasn't all in
one file. I have no idea if I did it correctly.

But anyway, I tested this with group sizes of 3 and 4 and in both cases
answers were found in the input! The group size of 4 took quite a few
iterations, and it definitely goes up quickly for larger group sizes.
Anyway, good learning experience.