Add Combinations len method and more process output info
This was kind of fun. I had looked up how to figure out how many
possible combinations of group_size there would be in a set of elements,
and came across the formula. I first tried implementing it naively with
all the factorials, which quickly overflowed usize. Then I realized I
could pretty easily reduce the factorial fraction to keep the numbers
smaller and it now works at least for our input size of 200 and group
sizes of 2, 3, 4, and 5.

I tried out the test with those sizes and solutions were found up
through 4, which was neat, but no solution was found for 5.