Day 4 part 2
I'm not sure how idiomatic this solution is or how memory efficient. I
wanted to be able to iterate over the validators rather than calling
them explicitly and &&ing together the results. This creates some
overhead, but it was fun to figure out. I also learned a few things
about how match expressions work and learned the value of always using
them when you're not sure you have a value, and not just calling unwrap.