Day 6 part 2
I really struggled with this part 2, trying to come up with an elegant
solution. I'm still not sure if I'm very happy with it. I think it's
largely getting used to the ways that Rust does things that I'm used to
doing in Ruby or Elixir. I'm probably doing things in a not very
idiomatic way for Rust.